Convert to NS2.0 implementation
We remove the dependency on gcmcore as a binary, which at the moment targets net7.0, limiting this library portability.

By selectively bringing sources as-is, we can target NS2.0 while keeping a compatible implementation that's still easy to maintain in sync with upstream.

By leveraging ILRepack, we can internalize all types from gcmcore and therefore maintain a slim API with just the relevant interfaces.

Fixes #93